Se agregaron metodos gets

+++ Models/ProductoVendido.cs
@@ -2,9 +2,9 @@
 {
     public class ProductoVendido
     {
-        public int Id { get; set; }
-        public int IdProducto { get; set; }
+        public long Id { get; set; }
+        public long IdProducto { get; set; }
         public int Stock { get; set; }
-        public int IdVenta { get; set; }
+        public long IdVenta { get; set; }
     }
 }
