+++ Models/Venta.cs
@@ -2,9 +2,9 @@
 {
     public class Venta
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Comentarios { get; set; }
-        public int IdUsuario { get; set; }
+        public long IdUsuario { get; set; }
 
     }
 }
